Add generateToken, verifyToken and decodeToken functions
